Add files via upload
This mini-project might appear useless or easy to finish. But to me, it gives value because it is a real-world related problem being that it deals with school systems. It aims to apply the skills and knowledge that I previously studied, I designed this program for academic grades, but this time all data is entered by the user via the input() function and is saved in a text file. This program is for recording academic grades, where the user is asked for his/her name and grades in some subjects, and then it calculates the grade point average and saves all that data.

The program first ask the user to enter the name of the file and then requests his/her data, starting with the name then the academic grades in some of the subjects specified in the program. Upon completion, the program saves all the data. It also calculates the student's academic average and saves it and prints it as well. The input() function is used once inside the loop, but the values that appear in it change constantly. This is because I used variables in this function, where after the first cycle in the loop, the value of the variable (m) to 'mark' and the value of item keeps changing every time. A cycle starting from the first item in the list of courses to the last item.